backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
package-mode = false
authors = [
    {name = "Nurgisa Saginbek",email = "nurgisasaginbek@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.121.0,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "openai (>=2.7.1,<3.0.0)",
    "elevenlabs (>=2.22.0,<3.0.0)",
    "alembic (>=1.17.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
